requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
random-username==1.0.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
